pyproject.toml
[tool.poetry]
name = "verbafloresumemaker"
version = "0.1.0"
description = ""
authors = ["Aditya Ghai <adityaghailbdrp1@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.3"
openai = "^1.45.0"
pypdf2 = "^3.0.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
